requirements.txt
streamlit>=1.8.1
streamlit-folium==0.6.15
pandas==1.3.4
numpy==1.21.4
Pillow==7.0.0
pyproj==3.4.0
haversine==2.7.0
Shapely==1.8.5
tqdm==4.62.3
geopandas==0.11.1
geojson==2.5.0
protobuf==3.20.*
requests==2.22.0
requests-html==0.10.0
